wts 파일 다운로드 인코딩 형식 변경, 함수 위치 수정
